change using geter and seter

+++ SchoolSystem/Entities/Teacher.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class Teacher {
     private String name;
     private String id;
-    private Short yearsOfExperience;
+    private Integer yearsOfExperience;
     private List<Subject> subjectsExpertiseList;
 
     public String getName() {
@@ -24,11 +24,11 @@ public class Teacher {
         this.id = id;
     }
 
-    public Short getYearsOfExperience() {
+    public Integer getYearsOfExperience() {
         return yearsOfExperience;
     }
 
-    public void setYearsOfExperience(Short yearsOfExperience) {
+    public void setYearsOfExperience(Integer yearsOfExperience) {
         this.yearsOfExperience = yearsOfExperience;
     }
